+++ lib/view/screens/goals_screen/goal_open.dart
@@ -3,6 +3,7 @@ import 'package:life_leaf/controller/goals_db_functions/goals_db_functions.dart'
 import 'package:life_leaf/model/goals_model/goals_main_model.dart';
 import 'package:life_leaf/model/goals_model/goals_model.dart';
 import 'package:life_leaf/view/screens/goals_screen/goals_edit.dart';
+import 'package:life_leaf/view/widgets/scaffold_messenger.dart';
 
 class GoalOpen extends StatefulWidget {
   final String title;
@@ -10,12 +11,15 @@ class GoalOpen extends StatefulWidget {
   final GoalsMainModel goal;
   final String mainGoalkey;
   final String stepkey;
+  final int isMarked;
   const GoalOpen(
       {super.key,
       required this.title,
       required this.stepss,
       required this.mainGoalkey,
-      required this.goal, required this.stepkey});
+      required this.goal,
+      required this.stepkey,
+      required this.isMarked});
 
   @override
   State<GoalOpen> createState() => _GoalOpenState();
@@ -26,7 +30,7 @@ class _GoalOpenState extends State<GoalOpen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: Color.fromARGB(255, 62, 62, 62),
+      backgroundColor: const Color.fromARGB(255, 49, 49, 49),
       appBar: AppBar(
         elevation: 0,
         backgroundColor: Colors.transparent,
@@ -38,8 +42,10 @@ class _GoalOpenState extends State<GoalOpen> {
         ),
         title: Text(
           widget.title,
-          style: TextStyle(
-              color: Color.fromARGB(255, 169, 249, 172), fontSize: 20),
+          style: const TextStyle(
+              color: Color.fromARGB(255, 227, 251, 92),
+              fontSize: 25,
+              fontWeight: FontWeight.w800),
         ),
         actions: [
           PopupMenuButton(
@@ -47,7 +53,7 @@ class _GoalOpenState extends State<GoalOpen> {
             iconSize: 30,
             itemBuilder: (context) => [
               PopupMenuItem(
-                child: ListTile(
+                child: const ListTile(
                   leading: Icon(Icons.edit),
                   title: Text('Edit'),
                 ),
@@ -61,18 +67,23 @@ class _GoalOpenState extends State<GoalOpen> {
                           mainGoalkey: widget.mainGoalkey,
                           goal: widget.goal,
                           stepkey: widget.stepkey,
+                          isMarked: widget.isMarked,
                         ),
                       ));
                 },
               ),
               PopupMenuItem(
-                child: ListTile(
+                child: const ListTile(
                   leading: Icon(Icons.delete),
                   title: Text('Delete'),
                 ),
-                onTap: () {
-                  GoalDb.deleteGoal(widget.mainGoalkey);
-                  Navigator.pop(context);
+                onTap: () async {
+                  bool confirmed = await showDeleteConfirmation(context);
+                  if (confirmed) {
+                    GoalDb.deleteGoal(widget.mainGoalkey);                                          // Perform delete operation
+                    // ignore: use_build_context_synchronously
+                    Navigator.pop(context);
+                  }
                 },
               ),
             ],
@@ -85,28 +96,30 @@ class _GoalOpenState extends State<GoalOpen> {
             itemCount: widget.stepss.length,
             itemBuilder: (context, index) {
               return Card(
+                  color: const Color.fromARGB(255, 238, 208, 155),
                   child: ListTile(
-                title: Text(widget.stepss[index].title),
-                subtitle: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(widget.stepss[index].description),
-                    Text(widget.stepss[index].targetDate ?? '')
-                  ],
-                ),
-                onLongPress: () {
-                  //Delete option
-                },
-                trailing: Checkbox(
-                  value: isTicked,
-                  activeColor: Colors.green,
-                  onChanged: (newBool) {
-                    setState(() {
-                      isTicked = !isTicked!;
-                    });
-                  },
-                ),
-              ));
+                    title: Text(
+                      widget.stepss[index].title,
+                      style: const TextStyle(
+                          fontWeight: FontWeight.w900,
+                          fontSize: 20,
+                          color: Color.fromARGB(255, 255, 255, 255)),
+                    ),
+                    subtitle: Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        Text(
+                          widget.stepss[index].description,
+                          style: const TextStyle(
+                              fontWeight: FontWeight.w600, fontSize: 18),
+                        ),
+                        Text(widget.stepss[index].targetDate ?? '')
+                      ],
+                    ),
+                    onLongPress: () {
+                      //Delete option
+                    },
+                  ));
             },
           )),
     );
